fix(model): pressure-test audit of the DeepLens core (crash, non-square, device, perf)
Squashed result of a full bug/efficiency pressure test of the DeepLens core.
Net changes:

Crash / dead-entry-point repairs:
- Repair crash-on-call paths in prism, cubic, optim, the PSFNet cluster,
  render_mono/render_rgbd, and the loss barrier (shape/depth/foc_dist handling).

Correctness:
- Fix non-square sensor/field W/H swaps across wave optics & PSF
  (FresnelDiffraction/FraunhoferDiffraction/gen_xy_grid/RayleighSommerfeld).
- init_device() no longer auto-selects MPS, so float64 wave optics run on
  Apple Silicon (falls back to CPU; pass device="mps" for the float32 path).
- material set_sellmeier_param sets dispersion="sellmeier"; Mirror.surf_dict
  stores d as a JSON-serializable float; calc_chief_ray_infinite uses each
  field's own depth; NURBS keeps the default dtype; vortex load_ckpt reads the
  "charge" key.

Performance:
- Cache rotation matrices in to_local/to_global_coord.
- Shrink the pupil-estimation ray fan (O(N^2) lstsq) from 1024 to 128 rays.
- Vectorize NURBS phi() (drop the per-point Python loop).

Per the owner's lean-code preference, defensive additions trialed during the
audit were deliberately NOT kept: phase-surface coefficients stay bare tensors
(not nn.Parameter); no extra NaN/empty guards (fresnel |f0|, eval 1/d_z, optim
loss_bound, psf_huygens zero-ray, PSFLoss +eps); no _cuda_is_available wrapper,
get_optimizer_params_manual path, or SPP_PUPIL constant. The off-axis
coherent-PSF spot-centering change was trialed and reverted (recenter still
crops at the geometric psf_center).

Adds regression tests: crash repairs, non-square wave optics, device selection,
remaining NaN guards, efficiency, and the minor-cleanup batch.

+++ 0_hello_diffraclens.py
@@ -21,6 +21,7 @@ Technical Paper:
         High Dynamic Range Imaging," CVPR 2020.
 """
 
+import torch
 from torchvision.io import read_image
 from torchvision.utils import save_image
 
@@ -32,7 +33,11 @@ from deeplens.imgsim import conv_psf
 # =====================================================================
 # Load a minimal diffractive lens (a single Fresnel DOE focusing at f0 = 50 mm,
 # one focal length in front of the sensor) from a JSON configuration file.
-lens = DiffractiveLens(filename="./datasets/lenses/diffraclens/fresnel.json")
+# Run in float64 to match the docstring's numerical-stability claim for the
+# wave-propagation step (DiffractiveLens defaults to float32 otherwise).
+lens = DiffractiveLens(
+    filename="./datasets/lenses/diffraclens/fresnel.json", dtype=torch.float64
+)
 
 # =====================================================================
 # PSF analysis
